003 End of day commit
March 28th 2017
1. Accordion works
2. Accordion gets fields
3. Page is centered
4. Changed Archive-ourimpact to taxonomy-impact.php
5. Changed Archive-ourimpact.scss to taxonomy-impact.scss
6. Delete archive-localfood.php
7. Delete archive-systemsthinking.php

What needs to be done:
1. Create breadcrumbs
2. Make marker arrow point to the page you are currently on
3. Add little arrow image showing that accordion can be opened
4. Centre category titles
5. Create and add the blue box that can dynamically change the
percentage of waste,  number of waste diverted and pounds diverted.
-and place it above the upcycling category and below the the composting
category however this is does not open and is not itself an accordion
and does not show on the other pages(ie. Local Food Procurement and
Systems Thinking.)
6. Show the Taxonomy Names in the order the comp has it (ie. Zero Waste,
Local Food Procurement, Systems Thinking)